+++ packages/plugin-compiler-alipay/src/templateProcessorToAlipay.ts
@@ -15,6 +15,9 @@ const TAG_SPECIFIC_PROP_NAME_MAPPINGS = {
     activeColor: 'active-color',
     'block-size': 'handle-size',
     'block-color': 'handle-color'
+  },
+  map: {
+    polygons: 'polygon'
   }
 }